Don't render all bookmarks at once
This change considerably improves rendering performance when one has a
lot of bookmarks stored.

The app now only renders the most recent 30 bookmarks by default. It
will then detect when the last item comes into view and load 30 more. It
will do this as long as there are more bookmarks to load. Keeping it
simple.

closes #68