Optimize "back to list" behavior with back button
This changes the button so that it replaces the current route when
clicking "back to list" after having added (or edited or deleted) a
webmark.

This considerably improves the behavior when a user saves a webmark from
the bookmarklet or the system's native share dialog (e.g. on Android).
Currently, if they're going back to the list, and then decide to go back
to the original page, they only go back to the add/edit form. With this
change, they go back immediately to the original page. Or, when using
the system's share function, the entire overlay browser window will be
closed.